Use immutable tuples for __all__
This is to address docstring linting warnings from pydocstyle
(and only fixed the top level files by hand as a proof of
principle - there are so many more it would be worth automating).

Also I am consistently using a multi-line layout even when
there is a single entry, and using double quotes.

Should close issue #607